backend/requirements.txt
blinker==1.8.2
click==8.1.8
Deprecated==1.2.18
flask==3.0.3
gunicorn==23.0.0
importlib-metadata==8.5.0
itsdangerous==2.2.0
jinja2==3.1.6
lxml==5.4.0
MarkupSafe==2.1.5
packaging==25.0
pillow==10.4.0
PyMuPDF==1.24.11
werkzeug==3.0.6
wrapt==1.17.2
zipp==3.20.2
